:sparkles: feat: mapX, mapY double 형으로 바꿈

+++ JejuDorang/src/main/java/JejuDorang/JejuDorang/tourspot/service/TourSpotService.java
@@ -50,8 +50,8 @@ public class TourSpotService {
                 String address = itemNode.path("addr1").asText();
                 String image = itemNode.path("firstimage").asText();
                 String image2 = itemNode.path("firstimage2").asText();
-                Long mapX = itemNode.path("mapx").asLong();
-                Long mapY = itemNode.path("mapy").asLong();
+                Double mapX = itemNode.path("mapx").asDouble();
+                Double mapY = itemNode.path("mapy").asDouble();
 
                 tourSpotResponseDtos.add(new TourSpotResponseDto(
                         title, address, image, image2, mapX, mapY));
